Copy template login_simulation.html from djangorestlogin
It is to make installation more convenient.  Now we are are installing
djangorestlogin with pip, but it is not easy to configure Django to use
templates from external libraries in portable fashion.  It is only a
single template with a few lines anyway.  We are coping it to templates
of the module common as the workaround.