dataset: Prevent changes in case of an error on commit
We were thinking that we are on the safe side by setting ATOMIC_REQUESTS
on the settings.  It was causing the requests to operate in a database
transaction, but when and exception happened and handled, the
transaction was not rolled back.  This commit reverts the setting, and
enabled atomic transaction context for dataset.commit.  We can use it
other place, we notice the same problem.