Don't filter servertypes by intern_ip
We were trying guess from the searched inter_ip whether the user desire
an host or a network.  This was added to not cause backwards
incompatibility after networks are introduced.  However, I accidentally
broke it with the last update, and nobody noticed.  Not treating
anything specially is surely a better design.