Rip out the serveradmin API function loader middleware
From now on every serveradmin app is responsible to load its own api
functions via their apps.py. Django middlewares are a way to manipulate
requests on their way in or out. The API middleware is misappropriating
the fact that the middleware will be loaded during the django startup
to import and register the api functions. Since django 1.9 the apps.py
can be used to run arbitary code when an app is loaded to achive this.
Django2 broke the current api middleware style that's why I'm making this
change now.