Split settings.py into default and user settings
The user won't care to change and maintain many of the settings in our
settings.py. Loading a set of defaults first and the user settings later
means less work for other devs and users after a setting was changed.

From now on users just need to create a local_settings.py with a
SECRET_KEY = 'SOME-SECRET' to get a basic serveradmin running.