Revert "Improve uniq ip constraints"
This reverts commit 963274a77ccc09ea43634224a53064dc5ffea23c.

These constraints try to limit values of server.intern_ip based on the
value of server.servertype.ip_addr_type.  Postgresql doesn't allow
constraints based on columns of foreign tables.  This check tried to be
smart by creating a function and marking it immutable.  Immutable
functions guarantee to always give the same output, given the same
input.  That allows postgresql to cache the result which is a
prerequisite for them to be used in constraints.

This promise is already somewhat of a lie as changing the servertypes
ip_addr_type would not make postgresql clean the cached results from
this function.  That was an acceptable trade off for us as we never
changed the ip_addr_type on an existing servertype anyway.  Neither did
we handle the resulting need to re-check all intern_ip values.

What we didn't anticipate was, that pg_{dump,restore} isn't smart
enough to order restoring the server and servertype rows before
checking the constraints.  If pg_restore inserts server rows, with
servertypes not yet inserted into the servertype table, the function
will fail and the row will not be inserted.  Hence the dumps can no
longer be restored.