acl: Fix vulnerability allowing the change of non-whitelisted attributes
The materialized objects we get in the _access_control() function
always include all attributes of the servertype, even if those
attributes were not touched in this commit.  Hence, on changed
objects, we compare all attribute values with their previous value.
If the values hasn't changed, we accept the attribute value even if
the attribute id is not on our whitelist.

At least since this module was flattened a year ago this code compared
the new attribute value against, again, the new attribute value.
Obviously this check always returns equality and allows the change.