servershell(fix): User types to fast for auto completion
Some users rely on typing very fast a part of a command and then hit
enter. With the delay of 100 ms they see to often command does not
exist. It is safe for command to reduce the delay to 10 ms because we
do not make any requests to the backend which could overload the backend
with requests.