fix(adminapi): Do not sent already expired request
The Serveradmin backend rejects all requests which have been created 17
seconds or earlier in the past or 17 seconds and later in the future to
avoid replay attacks but our timeout is set to 60 seconds and we retry a
request up to 3 times.

When retrying to send a request after 60 seconds of timeout the request
is expired for sure and we must recreate it. For other resaons such as
HTTPError, SSLError or URLError we don't know but when we recreate it we
are on the safe side.

The timeout of urlopen is used for all blocking operations so we can not
distinguish between connection attempt and a open running connection
between the client and the server. While a timeout greater 17 seconds
for request on connection attempt is useless a timeout for a long
running API call of 60 seconds makes sense. It makes therefore no sense
to reduce the timeout to 16 seconds or less.