feature: Provide server side DatasetObject
We provide already a server side Query implementation which allows executing
Serveradmin queries without doing an actual HTTP request including all
the overhead of authorization and authentication by just passing the
user or application directly.

This makes also the DatasetObject and its commit method available as
server side implemention.

This can be useful for many things but the intention here is to be able to
write proper unit tests and being able to commit data in a isolated test
database using the Django TransactionTestCase class.