Guess attribute data type and apply conversion
This code is also used in the Servershell when using the edit page to
transform the user input to the correct data type.

We want to allow people to set attribute values as string also when the
attribute type is not string e.g. host['intern_ip'] = '10.0.0.1' and
automatically convert it to the correct data type.