Defer some special attributes in queries
The way Serveradmin is implemented assumes some special attributes to
be always present. This is object_id, hostname, intern_ip and
servertype. No matter what query is executed, all of these attributes
are fetched, whether they are needed or not. Especially for intern_ip
this has a HUGE impact on the overall performance since it is parsed
into an IP object. If additional relation attributes are requested by
the user, this performance impact can quickly rise exponentially
because the relations can be many-to-one or even many-to-many.
Imagine a query that requests 1000 servers with a relation and each
server has multiple of those relations, the processing time
immediately skyrockets by an order of magnitude compared to a query
without any relations.
The second most impacting attribute but by far not as much impacting
is servertype because it is already a relation by itself.
This commit therefore defers the fetching of intern_ip and servertype
attributes in queries with relations by default.
The intern_ip attribute is also deferred when the servers are fetched
before resolving relations and other attributes, because intern_ip is
rarely ever needed, but still adds a lot of overhead in all requests.
The servertype attribute is not deferred there, because servertypes
must be known upfront, before resolving other attributes.
On my local machine this saves a good ~40% of processing time for
queries that involve relations. For queries without relations, this
does not improve performance much (maybe 5% max), but it does also
not make it worse. Which justifies the added deferred overhead.