Optimize the way history data is loaded
Prefetch all data in big single queries instead of fetching every
little piece of information in a separate query. For large histories
this improves performance by an order of magnitude. In my case it
reduces processing time from ~360s to ~32s, which is ~1100% faster,
for a history with more than 100k changes.
This is still not a perfect solution. It should be discussed whether
we want to change the data model properly or write custom queries.
I see a huge potential for further performance gains by doing that.