Prefetch all many relations instead of FULL JOIN
A lot of time is wasted when fetching many-to-one or many-to-many
relations because of Django's select_related() which does FULL JOINs.
There is another method prefetch_related() that is especially useful
for many relations as it makes separate queries and joins requested
objects later inside Python. The more relations an object has the
higher is the performance impact.
In my simple test cases I was able to reduce processing time by a
good ~45%. For nested queries this therefore improves the performance
by a lot.