Make type read-only for editing existing attributes
Changing the attribute type of existing attributes can be a dangerous
operation because we don't have a defined behaviour for the existing
values so it is not really supported.

Techincally one could make changes to attribute types for existing
objects by e.g. writing a script that dumps the values, changes the
attribute type and then restores the values but this required deep
knowledge and is error prone so it is best to deny it in the first
place.

For know we handle such cases manually in the database and hopefully
come up with support for changing attribute types in a safe manner in
future.