Restoring an object with "computed" attributes fails (#306)
There are attributes we call "computed" which don't have stored values
in the database. Their values are based on other attributes of the same
object or related objects.

Setting values for these attributes when creating an object fails with
an exception and makes no sense so we exclude them here.

One might argue to not store them for deleted objects in the first place
but the values might still be of interest e.g. VMs that ran on a
hypervisor.