Skip check for permission to change for related_via attrs
Attributes on servertypes which are configured to get their value from
another servertype (related_via) should be excluded from the permission
check to change them because the actual change takes place on the target
servertype and is also checked there.

Thus the check would be redundant and annoying because you have to
whitelist the attributes in two places in ACLs.