Allow changes if permitted by app or user
Change the current behavior so that a change must be permitted by the
applications ACLs or the users ACLs (if there is no application) but no
longer the application and the owner of the application.

This is the wanted behavior because we either commit changes from the
Servershell where we want to evaluate the permissions of the logged in
user or from the Adminapi where we want to evaluate the permissions of
the application. We do not care about the owner of the application for
the evaluation as applications should be stand alone.

The ownership is relevant for disabling all applications when a user
gets disabled and of course to keep track of who uses the application.

The old behavior has historical reasons that are no longer met.