Retry signing the request if too slow
There are cases where the cryptographic operation to sign the request
takes unusually long. One example is when using a hardware such as the
YubiKey which is busy signing other stuff.

The Serveradmin server is pretty strict and would reject these as expired
for security reasons.

To gracefully handle temporary congestion in the signing we log an error
and retry the operation up to Settings.tries.