Drop obsolete constraint on server table (#371)
This constraint ensured that they are no duplicate ip addresses per
servertype for objects of ip address type other than loadbalancer.

This is now ensured and checked in the clean method of the Server model.
This has the advantage that errors are raised as validation exceptions
properly displayed in the Servershell and adminapi as errors. The
constraints are swallowed as database exception.