Fix is_unique_ip logic
We allow Servers to have various inet attribute types which configure
behaviour of checking for IP address conflics. Type "loadbalancer" does
not have an uniqueness check, type "network" only checks within the same
servertype.

Ensure that the uniqueness of IP address is checked symmetrically:
"host" inet attribute type must not ensure its uniqueness against
"network" and "loadbalancer", because they don't check against "host".

For example it was possible to create a domain object for IP address of
a VM, but it was impossible to create a VM once a domain existed.

Improve the message shown to user in case of conflict.