Calculate supernets separately per address family
Supernet is calculated only for intern_ip and expected to yield a single
value for each Server. Most of our servers and loadbalancers are
dual-stack, that is they have IPv4 and IPv6 addresses. The algorithm
works fine for calculating route_network, as this is always a single
object with IPv4 and IPv6 addresses. But it fails for provider_network:
LB Pools might belong to a different provider_network for IPv4 and IPv6.
For example IPv4 might be leased from datacenter, while IPv6 is from our
own provider-independent prefix.

Provide option to obey target address family when calculating a
supernet. This allows for adding multiple supernet attributes e.g.
provider_network_ipv4 and provider_network_ipv6.